% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_GC_plot.R
\name{abundance_GC_plot}
\alias{abundance_GC_plot}
\title{Plot the relationship between GC content and feature abundance.}
\usage{
abundance_GC_plot(sep, bsgenome, fragment_length = 100,
  save_pdf_prefix = NULL, combine_replicates = FALSE)
}
\arguments{
\item{sep}{a \code{summarizedExomePeak} object.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence.}

\item{fragment_length}{the expected fragment length of the sequencing library.
The widths of the row features for quantifying GC content will be re-sized into the fragment length if it falls bellow it.}

\item{save_pdf_prefix}{a character, if provided, a pdf file with the given name will be saved under the current directory.}

\item{combine_replicates}{a logical indicating whether to pool the replicates in the local regression fit; default FALSE.}
}
\value{
a ggplot object
}
\description{
This function plot the local regression curves of the normalized feature abundance (reads count) against the local GC content levels.
}
\details{
The read abundances of both the control and the modification site regions are plotted,
the read counts are normalized using the following method:

 normalized feature abundance = ( ( read_count / size_factor) / region_length ) * 500

 By default, it will use the sequencing depth size factor defined in \code{summarizedExomePeak},
 if the size factor is not found, new size factors will be estimated using the default method in \code{estimate_size_factors}.
}
