% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_bg.R
\name{annot_bg}
\alias{annot_bg}
\title{Find the background of the user provided annotation.}
\usage{
annot_bg(annot, txdb, cut_off_width = 1e+05, cut_off_num = 2000,
  drop_overlapped_genes = TRUE, m6Aseq_background = TRUE,
  control_width = 50)
}
\arguments{
\item{annot}{A \code{GRanges} object of user provided annotation (names are neccessary for the index of the splitting).}

\item{txdb}{A \code{TxDb} object that define the transcript annotation.}

\item{cut_off_width}{A non-negative integer indicate the least total width of the disjoint exons used as the background; Default 1e5.}

\item{cut_off_num}{A non-negative integer indicate the leat total number of the disjoint exons used as the background; Default 2000.}

\item{drop_overlapped_genes}{A logical indicating whether to discard the overlapping genes; Default TRUE.}

\item{m6Aseq_background}{A logical value, TRUE if the region of 5'UTR and long exons of the transcripts should be dropped in control region; Default TRUE.}

\item{distance_5p}{A numeric value of the length of the transcript starting region; default 200.}
}
\value{
A \code{GRangesList} object.
The first portion is the exons regions that is not overlapped with \code{annoation}.

If the resulting ranges have less number and width compared with what defined in \code{cut_off_width} or \code{cut_off_num},
the exon regions of txdb will be returned as the background.

The second portion is the restructed user provided annotation with gene id annotated.
}
