% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks_with_DESeq.R
\name{call_peaks_with_DESeq}
\alias{call_peaks_with_DESeq}
\title{Statistical Inference with DESeq package based on the provided reads count for exomic bins.}
\usage{
call_peaks_with_DESeq(SE_bins, count_cutoff = 10, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0, txdb,
  drop_overlapped_genes = TRUE)
}
\arguments{
\item{SE_bins}{a \code{SummarizedExperiment} object. The meta-data collumn should contain the design information of IP & input and treated & control.}

\item{count_cutoff}{an integer value indicating the cutoff of the sum of reads count in a window, inference is only performed on the windows with read count bigger than the cutoff. Default value is 10.}

\item{p_cutoff}{a numeric value of the p value cutoff used in DESeq inference.}

\item{p_adj_cutoff}{a numeric value of the adjusted p value cutoff used in DESeq inference; if provided, the value of \code{p_cutoff} will be ignored.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{txdb}{the txdb object that is necessary for the calculation of the merge of the peaks.}

\item{drop_overlapped_genes}{A logical indicating whether the overlapping genes were dropped.}
}
\value{
This function will return a list of \code{GRangesList} object storing peaks for both methylation and control results,
}
\description{
\code{call_peaks_with_DESeq} conduct inference on every exome bins using negative binomial model,
the significant bins will be the merged into peaks.
}
\details{
\code{call_peaks_with_DESeq} will performe exome level peak calling using DESeq model,
the significant bins in both directions (IP > input and input > IP) are merged separately.
The final result will contains both the methylation peaks and control peaks.

If the experimental design contains additional factor of treatment,
the treated group will be tested independently of the control group.
The final peak region will be the reduced union of the peaks from 2 groups.
}
