% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks_with_DESeq2.R
\name{call_peaks_with_DESeq2}
\alias{call_peaks_with_DESeq2}
\title{Statistical Inference with DESeq package based on the provided reads count for exomic bins.}
\usage{
call_peaks_with_DESeq2(SE_bins, count_cutoff = 5, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0, txdb,
  drop_overlapped_genes = TRUE)
}
\arguments{
\item{SE_bins}{a \code{SummarizedExperiment} object. The meta-data collumn should contain the design information of IP/input + treated/control.}

\item{count_cutoff}{an integer value indicating the cutoff of the mean of reads count in a row, inference is only performed on the windows with the row average read count bigger than the cutoff. Default value is 5.}

\item{p_cutoff}{a numeric value of the p value cutoff used in DESeq inference.}

\item{p_adj_cutoff}{a numeric value of the adjusted p value cutoff used in DESeq2 inference; if provided, the value of \code{p_cutoff} will be ignored; Default = 0.05.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{txdb}{the txdb object that is necessary for the calculation of the merge of the peaks.}

\item{drop_overlapped_genes}{A logical indicating whether the overlapping genes were dropped.}
}
\value{
This function will return a list of \code{GRangesList} object storing peaks for both methylation and control.
}
\description{
\code{call_peaks_with_DESeq2} conduct inference on every exome bins using negative binomial model,
the significant bins will be the merged into peaks.
}
\details{
\code{call_peaks_with_DESeq2} will performe exome level peak calling using DESeq2 model,

The significant bins will be merged into methylation peaks.

The insignificant bins (pass the row means filtering) will also be merged into control peaks.
}
