% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consDESeq2_DM.R
\name{consDESeq2_DM}
\alias{consDESeq2_DM}
\title{Calculate the index for across sample differential modification consistency in a DESeqDataSet object}
\usage{
consDESeq2_DM(dds, consistent_log2FC_cutoff = 1, p = 0.7,
  alpha = 0.05)
}
\arguments{
\item{dds}{a DESeqDataSet object.}

\item{consistent_log2FC_cutoff}{a numeric for the cutoff of differential modification log2 fold changes; default = 1.}

\item{p}{the binomial parameter p used by the binomial test; default = 0.7.}

\item{alpha}{the significant level of the binomial test; default = 0.05.}
}
\value{
a logical index for the consistently modified rows in the DESeqDataSet.
}
\description{
Calculate the index for across sample differential modification consistency in a DESeqDataSet object
}
\details{
The minimum consistent number cutcoff is defined by 1-alpha quantile of a binomial distribution with probability of success = p, and number of trials = number of possible pairs between replicates.
This is equivalent to the rejection of an exact binomial test with null hypothesis of p < 0.8 and N = number of possible pairs.
}
