% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consDESeq2_M.R
\name{consDESeq2_M}
\alias{consDESeq2_M}
\title{Calculate the index for across sample modification consistency in a DESeqDataSet object}
\usage{
consDESeq2_M(dds, consistent_log2FC_cutoff = 1,
  consistent_fdr_cutoff = 0.05, p0 = 0.8, alpha = 0.05)
}
\arguments{
\item{dds}{a DESeqDataSet object.}

\item{consistent_log2FC_cutoff}{a \code{numeric} for the modification log2 fold changes cutoff in the peak consisency calculation; default = 1.}

\item{consistent_fdr_cutoff}{a \code{numeric} for the BH adjusted C-test p values cutoff in the peak consistency calculation; default { = 0.05}. Check \code{\link{ctest}}.}

\item{p0}{a \code{numeric} for the binomial proportion parameter used in the consistent peak filter; default \code{= 0.8}.

For a peak to be consistently methylated, the minimum number of significant enriched replicate pairs is defined as the 1 - alpha quantile of a binomial distribution with p = p0 and N = number of possible pairs between replicates.

The consistency defined in this way is equivalent to the rejection of an exact binomial test with null hypothesis of p < p0 and N = replicates number of IP * replicates number of input.}

\item{alpha}{a \code{numeric} for the binomial quantile used in the consitent peak filter; default\code{ = 0.05}.}
}
\value{
a logical index for the consistently modified rows in the DESeqDataSet.
}
\description{
Calculate the index for across sample modification consistency in a DESeqDataSet object
}
\details{
The minimum consistent number cutcoff is defined by 1-alpha quantile of a binomial distribution with probability of success = p, and number of trials = number of possible pairs between replicates.
This is equivalent to the rejection of an exact binomial test with null hypothesis of p < 0.8 and N = number of possible pairs.
}
