% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertTxDb}
\alias{convertTxDb}
\title{Transformation of Txdb object}
\usage{
convertTxDb(txdb, type = c("full_transcript", "whole_genome"))
}
\arguments{
\item{txdb}{a TxDb object of standard transcript annotation.}

\item{type}{the type of output, should be one of c("full_tx","whole_genome").}
}
\value{
a TxDb object.
}
\description{
Converting TxDb object into full transcript or whole genome modes.
}
\keyword{internal}
