% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructors.R
\name{Corgi}
\alias{Corgi}
\title{Dog construction}
\usage{
Corgi(personality = c("joyful", "neurotic", "cruel", "merciful"),
  weight = 0, length = 0, speed = 0, beauty = 0, satisfaction = 0,
  fur_amount = 0)
}
\arguments{
\item{personality}{can be one of the "joyful", "neurotic", "cruel", and "merciful".}
}
\description{
This function defines the constructors for dog class.
}
