% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_deseq.R
\name{decision_deseq}
\alias{decision_deseq}
\title{Calculate the decision table for a DESeq result.}
\usage{
decision_deseq(res, log2FC_cut, p_cut, padj_cut, min_mod)
}
\arguments{
\item{res}{A \code{DESeqResults} or similar object that contains the result statistics for either modification or differential modification.}

\item{log2FC_cut}{The log2 fold change cutoff of the inference result, default setting is 0.}

\item{p_cut}{A numeric value between 0 to 1, indicating the p value cut off of the Wald test defined by DESeq, it will be neglected if \code{padj_cut} is not NULL.}

\item{padj_cut}{A numeric value between 0 to 1, indicating the fdr cut off of the Wald test defined by DESeq.}

\item{min_mod}{Minimum number of features returned, when this is smaller than the cut-off results, additional features are called by the order of p values.}
}
\value{
A \code{data.frame} object indicating the column and cut-off value used for desicion, also it includes the number of positive sites in both directions based on the decision.
}
\description{
\code{Decision_dsresult} is an internal function used to summary the cut-off and the number of positive results used for DESeq result..
}
