% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_deseq2.R
\name{decision_deseq2}
\alias{decision_deseq2}
\title{Calculate the decision table for a DESeq2 result.}
\usage{
decision_deseq2(Inf_RES, log2FC_cut = 0, P_cut = 0.05,
  Padj_cut = NULL, Min_mod = 1000, Exp_dir = c("hyper", "hypo",
  "both"))
}
\arguments{
\item{Inf_RES}{A \code{DESeqResults} or similar object that contains the result statistics for either methylation or differential methylation.}

\item{log2FC_cut}{The log2 fold change cutoff of the inference result, default setting is 0.}

\item{P_cut}{A numeric value between 0 to 1, indicating the p value cut off of the Wald test defined by DESeq2 (or defined by QNB), it will be neglected if \code{Padj_cut} is not NULL.}

\item{Padj_cut}{A numeric value between 0 to 1, indicating the fdr cut off of the Wald test defined by DESeq2 (or defined by QNB).}

\item{Min_mod}{Minimum number of features returned, when this is smaller than the cut-off results, additional features are called by the order of p values.}

\item{Exp_dir}{This parameter is filled when making decisions on differential methylation, it could be "hyper", "hypo", and "both".}
}
\value{
A \code{data.frame} object indicating the column and cut-off value used for desicion, also it includes the number of positive sites in both directions based on the decision.
}
\description{
\code{Decision_dsresult} is an internal function used to summary the cut-off and the number of positive results used for DESeq2 result..
}
