% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{DESeq2Results-methods}
\alias{DESeq2Results-methods}
\alias{DESeq2Results,SummarizedExomePeak-method}
\alias{DESeq2Results}
\alias{DESeq2Results<-,SummarizedExomePeak-method}
\alias{DESeq2Results<-}
\title{Method DESeq2Results}
\usage{
DESeq2Results(x1)

DESeq2Results(x2) <- value

\S4method{DESeq2Results}{SummarizedExomePeak}(x1)

\S4method{DESeq2Results}{SummarizedExomePeak}(x2) <- value
}
\arguments{
\item{x1}{A \code{data.frame} object.}

\item{x2}{A \code{SummarizedExomePeak} object.}

\item{value}{a \code{data.frame} object for the DESeq2 Results.}
}
\value{
A \code{data.frame} object for the DESeq2 Results.
}
\description{
Method DESeq2Results

Method DESeq2Results<-

Accessor to the slot \code{DESeq2Results} in class \code{SummarizedExomePeak}.

Accessor to the slot \code{DESeq2Results} in class \code{SummarizedExomePeak}.
}
\examples{

f1 = system.file("extdata", "sep_ex_mod.rds", package="exomePeak2")

sep <- readRDS(f1)

head(DESeq2Results(sep))

}
