% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{DESeq2Results}
\alias{DESeq2Results}
\title{Accessor to the slot \code{DESeq2Results} in class \code{SummarizedExomePeak}.}
\usage{
DESeq2Results(x)

\S4method{DESeq2Results}{SummarizedExomePeak}(x)
}
\arguments{
\item{x}{A \code{SummarizedExomePeak} object.}
}
\value{
A data.frame.
}
\description{
Accessor to the slot \code{DESeq2Results} in class \code{SummarizedExomePeak}.
}
