% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_inference.R
\name{DESeq_inference}
\alias{DESeq_inference}
\title{Statistical Inference with DESeq on IP over input ratio.}
\usage{
DESeq_inference(count_assay, design_IP, p_cutoff = NULL,
  p_adj_cutoff = 0.05, count_cutoff = 10, logFC_meth = 0,
  logFC_control = 0, control_size = NULL)
}
\arguments{
\item{count_assay}{a \code{matrix} of read count. Rows of the matrix represent the peaks, and collumns represent the samples.}

\item{design_IP}{a character vector with values of "IP" and "input", which will indicate the design of MeRIP-Seq experiment.}

\item{p_cutoff}{a numeric value of the p value cutoff used in DESeq inference.}

\item{p_adj_cutoff}{a numeric value of the adjusted p value cutoff used in DESeq inference; if provided, values in \code{p_cutoff} will be ignored.}

\item{count_cutoff}{an integer value indicating the cutoff of the sum of reads count in a window, inference is only performed on the windows with read count bigger than the cutoff. Default value is 10.}

\item{logFC_meth}{a non negative numeric value of the log2 fold change cutoff used in DESeq inferene for methylated peaks (IP > input).}

\item{logFC_control}{a non positive numeric value of the log2 fold change cutoff used in DESeq inferene for control peaks (IP < input).}

\item{control_size}{an integer value indicating the number of randomly selected control peaks returned, by default, all the other peaks after the count cutoff are used as the control.}
}
\value{
a list of the index for the significant methylated peaks (IP > input) and control peaks (peaks other than methylation peaks).
}
\description{
\code{DESeq_inference} conduct inference on log2 fold changes of IP over input using the negative binomial model intruduced in DESeq.
}
