% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disj_background.R
\name{disj_background}
\alias{disj_background}
\title{Find the background of the user provided modification.}
\usage{
disj_background(
  mod_gr,
  txdb,
  background_bins = NULL,
  background_types = c("Gaussian_mixture", "m6Aseq_prior", "manual", "all"),
  control_width = 50,
  rename_mod = FALSE
)
}
\arguments{
\item{mod_gr}{A \code{GRanges} object of user provided modification (names are neccessary for the index of the splitting).}

\item{txdb}{A \code{TxDb} object that define the transcript modification.}

\item{background_bins}{A Granges object for background bins.}

\item{background_types}{A logical value, TRUE if the region of 5'UTR and long exons of the transcripts should be dropped in control region; Default TRUE.}

\item{control_width}{A integer for the minimum width of the control region returned; default 50.}

\item{rename_mod}{Whether to rename the returned modification sites, default = FALSE.}
}
\value{
A \code{GRangesList} object.
The first portion is the exons regions that is not overlapped with \code{annoation}.

If the resulting ranges have less number and width compared with what defined in \code{cut_off_num},
the exon regions of txdb will be returned as the background.

The second portion is the reconstructed provided modification with gene id mod_grated.
}
\description{
Find the background of the user provided modification.
}
\keyword{internal}
