% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_size_factors.R
\name{estimate_size_factors}
\alias{estimate_size_factors}
\title{Estimation of the size factors used in peaks quantification.}
\usage{
estimate_size_factors(sep, from = c("Control", "Methylation", "Both"), ...)
}
\arguments{
\item{from}{Determine which group the size factor is estimated from, can be one of the "Control", "Methylation", and "Both".

By default, the size factors are estimated from the merged control peaks, that correspond to the exonic regions other than the methylated peaks.}

\item{...}{inherited from \code{\link{estimateSizeFactorsForMatrix}}.}

\item{design_ip}{an optional logical vector indicating the design of IP and input, with TRUE represents for IP.}
}
\value{
This function will return a summarizedExomePeak object containing newly estimated sample wised size factors.
}
\description{
\code{estimate_size_factors} estimate the sample wised size factors on exomic peaks,
by default, it will use the median of the ratios to the row geometric means defined in package DESeq2.
}
\details{
The function takes the input of a summarizedExomePeak object, and it estimates the size factors based on the count information contained within it.
}
