% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exome_bins_from_txdb.R
\name{exome_bins_from_txdb}
\alias{exome_bins_from_txdb}
\title{extract exome bins for peak calling given a txdb object}
\usage{
exome_bins_from_txdb(txdb, window_size = 25, step_size = 25)
}
\arguments{
\item{txdb}{A \code{txdb} object.}

\item{window_size}{An integer valued number of the width of the sliding windows or bins.}

\item{step_size}{An integer valued number of the width of the bin steps.}
}
\value{
A \code{GRanges} object of exonic bins with the names corresponding to the indexes of bins.
A metadata colomn named gene_id is attached to indicate its gene ID, which is provided by the txdb object.
The gene IDs are divided into multiple ones if the gene contains exons that belong to different chromosomes and strands.
}
\description{
extract exome bins for peak calling given a txdb object
}
\keyword{internal}
