% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{exomePeak2Results-methods}
\alias{exomePeak2Results-methods}
\alias{exomePeak2Results,SummarizedExomePeak-method}
\alias{exomePeak2Results}
\alias{exomePeak2Results<-,SummarizedExomePeak-method}
\alias{exomePeak2Results<-}
\title{Method exomePeak2Results}
\usage{
exomePeak2Results(x1)

exomePeak2Results(x2) <- value

\S4method{exomePeak2Results}{SummarizedExomePeak}(x1)

\S4method{exomePeak2Results}{SummarizedExomePeak}(x2) <- value
}
\arguments{
\item{x1}{A \code{data.frame} object.}

\item{x2}{A \code{SummarizedExomePeak} object.}

\item{value}{a \code{data.frame} object for the DESeq2 Results.}
}
\value{
A \code{data.frame} object for the DESeq2 Results.
}
\description{
Method exomePeak2Results

Method exomePeak2Results<-

Accessor to the slot \code{exomePeak2Results} in class \code{SummarizedExomePeak}.

Accessor to the slot \code{exomePeak2Results} in class \code{SummarizedExomePeak}.
}
\examples{

f1 = system.file("extdata", "sep_ex_mod.rds", package="exomePeak2")

sep <- readRDS(f1)

head(exomePeak2Results(sep))

}
