% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exomePeakCalling.R
\docType{methods}
\name{exomePeakCalling}
\alias{exomePeakCalling}
\alias{exomePeakCalling}
\title{Exome peak calling on MeRIP-seq datasets while considering the biological variabilities.}
\usage{
exomePeakCalling(merip_bams = NULL, txdb = NULL, fragment_length = 100,
  binding_length = 25, step_length = binding_length, count_cutoff = 10,
  p_cutoff = NULL, p_adj_cutoff = 0.05, logFC_cutoff = 0,
  drop_overlapped_genes = TRUE, parallel = FALSE, mod_annotation = NULL,
  background = NULL)

\S4method{exomePeakCalling}{MeripBamFileList}(merip_bams = NULL,
  txdb = "hg19", fragment_length = 100, binding_length = 25,
  step_length = binding_length, count_cutoff = 5, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0, drop_overlapped_genes = TRUE,
  parallel = FALSE, mod_annotation = NULL, background = NULL)
}
\arguments{
\item{merip_bams}{a \code{MeripBamFileList} object.}

\item{txdb}{a txdb object, it could be a single character string that can be recognized by \code{\link{makeTxDbFromUCSC}}; Default "hg19".}

\item{fragment_length}{a positive integer of the expected fragment length in bp; default 100.}

\item{binding_length}{a positive integer of the antibody binding length in IP samples; default 25.}

\item{step_length}{a positive integer of the shift size of the sliding window; default is the binding length.}

\item{count_cutoff}{a non negative integer value of the average reads count per window used in peak calling, default to be 5.}

\item{p_cutoff}{a value of the p value cut-off used in peak calling.}

\item{p_adj_cutoff}{a value of the adjusted p value cutoff used in DESeq inference; default 0.05.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{drop_overlapped_genes}{a logical indicating whether the bins on overlapping genes are dropped or not, default to be TRUE.}

\item{parallel}{a logical indicating whether to use parallel computation, consider this if your computer has more than 16GB RAM.}

\item{mod_annotation}{a \code{GRanges} object for user provided single based RNA modification annotation. If provided, the peak calling step will be skipped.
Reads count will be performed using the provided annotation flanked by length of floor(fragment_length - binding_length/2).

The background regions used in this senario will be the disjoint exon regions of the flanked provided sites.}

\item{background}{a \code{GRanges} or \code{GRangesList} object for user provided background control regions on the genome.}
}
\value{
This function will return a \code{SummarizedExomePeak} object storing the ranges and reads counts information of the called peaks.
}
\description{
\code{exomePeakCalling} call RNA methylation peaks on exome regions with statistical tests that account for the biological variabilities between samples.
}
\details{
The function conduct exome level peak calling based on the read alignment results and the transcript annotations.
}
\examples{


}
\seealso{
\code{\link{exomePeakCalling}}
}
