% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonPlot.R
\name{ExonPlot}
\alias{ExonPlot}
\title{plot the distribution for the length of the overlapped exons.}
\usage{
ExonPlot(gfeatures, txdb, save_pdf_prefix = NULL)
}
\arguments{
\item{gfeatures}{a list of GRanges or GRangesList.}

\item{txdb}{a txdb object.}

\item{save_pdf_prefix}{provided to save a pdf file.}
}
