% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exportResults.R
\docType{methods}
\name{exportResults}
\alias{exportResults}
\alias{exportResults}
\title{Export the modification / differential modification sites}
\usage{
exportResults(sep, format = c("txt", "BED", "RDS"),
  save_dir = "exomepeaks_result", cut_off_pvalue = NULL,
  cut_off_padj = 0.05, cut_off_log2FC = 0, min_num_of_positive = 30,
  expected_direction = "both", inhibit_filter = FALSE,
  table_style = c("bed", "granges"))

\S4method{exportResults}{SummarizedExomePeak}(sep, format = c("tsv", "BED",
  "RDS"), save_dir = "exomePeak2_output", cut_off_pvalue = NULL,
  cut_off_padj = 0.2, cut_off_log2FC = 0, min_num_of_positive = 30,
  expected_direction = "both", inhibit_filter = FALSE,
  table_style = c("bed", "granges"))
}
\arguments{
\item{sep}{A SummarizedExomePeak object.}

\item{format}{A character string of one of the c("tsv", "BED", "RDS"), indicating the exported format.

- The choice tsv will save a tab separated values (tsv) file with the ranges and test information.

- The choice BED will save a BEDGraph file with the score column being the -log2 adjusted p value.

- The choice RDS will save a Rdata of the SummarizedExperiment object which will additional include a comprehensive summary of the count and the design information (recommended).}

\item{save_dir}{The name of the file being saved; Default "exomePeak2_output".}

\item{cut_off_pvalue}{A number between 0 and 1 indicate the p value cutoff in the exported result; Default NULL.}

\item{cut_off_padj}{A number between 0 and 1 indicate the adjusted p value cutoff in the exported result; Default 0.2.}

\item{cut_off_log2FC}{A non negative number indicating the log2 fold change cutoff of the exported result,

only sites with log2 IP/input fold change bigger than this value are kept; Default 0.

For differential modification analysis, the absolute value of the log2 Odds ratio will be filtered by \code{cut_off_log2FC}.}

\item{min_num_of_positive}{The minimum number of reported sites.
If the sites is filtered less than this number by its p values or effect sizes,
more sites will be reported by the order of the p value until it reaches this number.}

\item{expected_direction}{The expected differential modification direction, could be "hyper", "hypo", or "both".
This argument is useful when the treated group involves the perturbation of a writer or eraser protein for the modification; Default "both".}

\item{inhibit_filter}{Remove all the filters on the result, this is desired for user provided modification annotation; Default FALSE.}

\item{table_style}{Determine the style of the tsv table being exported, could be one of "bed" and "granges", the later would index the site containing multiple ranges with an id.}
}
\value{
a file containing the modification site genomic location, gene ids, statistics, and effect sizes will be created under the current working directory.
}
