% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flank_on_exons.R
\name{flank_on_exons}
\alias{flank_on_exons}
\title{Flank ranges on exon coordinate.}
\usage{
flank_on_exons(grl, flank_length, txdb, drop_overlapped_genes = TRUE,
  index_flank = TRUE)
}
\arguments{
\item{grl}{a GRangesList object which is the target of the flanking.}

\item{flank_length}{the length of the flanking regions (on both left and right).}

\item{txdb}{the TxDb object for the transcript annotation.}

\item{drop_overlapped_genes}{whether to drop the overlapping genes on the txdb object; default TRUE.}

\item{index_flank}{whether to store the flanking regions seperately; default TRUE.}
}
\value{
A GRanges object with the extented ends on exons.
The flanking regions and the binding regions are indexed by the names of the returned GRangesList.
The original input GRangesList will be dropped if it is not mapped to the exonic regions.
}
\description{
This function provides extention of the ends of the GRangesList on transcript coordinate or exons.
}
