% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content_over_grl.R
\name{GC_content_over_grl}
\alias{GC_content_over_grl}
\title{Retrieve GC content level from genome over regions defined by a GRanges or GRangesList object.}
\usage{
GC_content_over_grl(grl, bsgenome, fragment_length = 100)
}
\arguments{
\item{grl}{a GRanges or GRangesList object.}

\item{bsgenome}{a BSgenome object.}

\item{fragment_length}{an integer of the fragment length.}
}
\value{
a list contains 2 elements:

1. GC_content: the GC content of each GRanges / GRangesList element
2. Indx_length: the calculated total feature length used to quantify reads.
}
\description{
if the GRanges contain regions less than fragment_length, it will be re-sized to fragment_length by fixing at center.
This is currently not supported for user provided GRangesList annotations.
}
