% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_normalization.R
\name{GC_normalization}
\alias{GC_normalization}
\title{Estimation of the GC content normalization factors.}
\usage{
GC_normalization(sep, bsgenome = "hg19", feature = c("background", "all"),
  qtnorm = FALSE, fragment_length = 100, glm_offset = TRUE)
}
\arguments{
\item{sep}{a \code{summarizedExomePeak} object.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence, or it could be the name of this reference genome specified in a way that is accepted by the getBSgenome function defined in the BSgenome software package}

\item{feature}{the features used in the GC effect estimation, can be "background" and "all".
If "all" is choosed, the GC effect function will be estimated using both the methylated and the background region,
this choice will force the resulting modification signals independent of GC content.
However,it is possibably biased if the modification is biologically dependent on GC content;
Default "background".}

\item{qtnorm}{a logical indicating wheather perform quantile normalization after the GC effect estimation.
Quantile normalization can correct systematic effect in RNA-seq count data.
However, it may result in errors due to the biological difference between the distributions of IP and input.}

\item{fragment_length}{the expected fragment length of the sequencing library.
The widths of the row features for quantifying GC content will be re-sized into the fragment length if it falls bellow it.}

\item{glm_offset}{wheather return the generalized linear model normalization off-set; default TRUE.
The standard offset is useful when visualizing the corrected feature abundency.}
}
\value{
a \code{summarizedExomePeak} object containing the feature specific size factors.
}
\description{
This function estimates the feature specific size factors in order to correct the GC content artifacts,
the GC content biases can contributed to the following sources of variabilities in MeRIP-seq data:

1. Technical variablities between replicates.

2. Batch effects between different laboratories.

Using the default option, the GC normalization estimates the dependency between reads count and GC content only on the not methylated region,
This strategy can avoid the quantification biases for the modifications that biologically favor extream GC content.

The GC normalization can results in an improvement of accuracy for most published m6A-seq data,
and it is particullarly recommended if you want to compare the data between different laboratory conditions.
}
\details{
By default, the sequencing depth factor used is the column size factors obtained by \code{\link{estimate_size_factors}},
if the column size factors are not provided, it will be estimated using the default method defined in \code{\link{estimate_size_factors}}.
}
