% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{GCsizeFactors-methods}
\alias{GCsizeFactors-methods}
\alias{GCsizeFactors,SummarizedExomePeak-method}
\alias{GCsizeFactors}
\alias{GCsizeFactors<-,SummarizedExomePeak-method}
\alias{GCsizeFactors<-}
\title{Method GCsizeFactors}
\usage{
GCsizeFactors(x1)

GCsizeFactors(x2) <- value

\S4method{GCsizeFactors}{SummarizedExomePeak}(x1)

\S4method{GCsizeFactors}{SummarizedExomePeak}(x2) <- value
}
\arguments{
\item{x1}{A \code{SummarizedExomePeak} object.}

\item{x2}{A \code{SummarizedExomePeak} object.}

\item{value}{A \code{matrix} object.}
}
\description{
Method GCsizeFactors

Method GCsizeFactors<-

Accessor to the slot \code{GCsizeFactors} in class \code{SummarizedExomePeak}.

Accessor to the slot \code{GCsizeFactors} in class \code{SummarizedExomePeak}.
}
