% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_dm.R
\name{glm_dm}
\alias{glm_dm}
\title{Quantification and inference of the RNA differential methylation values based on the generalized linear models of negative binomial distribution.}
\usage{
glm_dm(sep, shrinkage_method = c("apeglm", "ashr"), ...)
}
\arguments{
\item{sep}{is a summarizedExomePeak object.}

\item{shrinkage_method}{a character indicating the method for emperical bayes shrinkage, can be one in "apeglm" and "ashr".
Please check \code{\link{lfcShrink}} for more information.}

\item{...}{inherited arguments from \code{\link{DESeq}}}
}
\description{
This function conducts a second round of RNA differential methylation inference based on an interactive generalized linear model of negative binomial distribution.

The differential methylation analysis is performed using the following design:

log2(Q) = intercept + I(Treatment) + I(IP) + I(IP):I(Treatment).

The statistics returned is calculated based on the coefficient estimate of the interactive term I(IP):I(Treated).

The resulting RNA differential methyltion level is quantified in form of the log2 Odds ratio; i.e. log2(IP to input ratio in Treatment / IP to input ratio in Control).

By default, the final returned log2 Odds ratio estimate will undergoes emperical Bayes shrinkage using a Couchey prior, which is defined in the package \package{apeglm}.
}
