% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_inference.R
\name{GLM_inference}
\alias{GLM_inference}
\title{Statistical Inference with DESeq2 on IP over input fold change.}
\usage{
GLM_inference(SE_bins, glm_type = c("poisson", "NB", "DESeq2"),
  p_cutoff = NULL, p_adj_cutoff = 0.05, count_cutoff = 5,
  logFC_mod = 0, min_mod_number = floor(nrow(SE_bins) * 0.001))
}
\arguments{
\item{SE_bins}{a \code{SummarizedExperiment} of read count. It should contains a colData with column named design_IP,
which is a character vector with values of "IP" and "input". The column helps to index the design of MeRIP-Seq experiment.}

\item{glm_type}{a character, which can be one of  the "poisson", "NB", and "DESeq2". This argument specify the type of generalized linear model used in peak calling; Default to be "poisson".
The DESeq2 method is only recommended for high power experiments with more than 3 biological replicates for both IP and input.}

\item{p_cutoff}{a numeric value of the p value cutoff used in DESeq inference.}

\item{p_adj_cutoff}{a numeric value of the adjusted p value cutoff used in DESeq2 inference; if provided, values in \code{p_cutoff} will be ignored.}

\item{count_cutoff}{an integer value indicating the cutoff of the mean of reads count in a row, inference is only performed on the windows with read count bigger than the cutoff. Default value is 10.}

\item{logFC_mod}{a non negative numeric value of the log2 fold change cutoff used in DESeq inferene for modification containing peaks (IP > input).}

\item{min_mod_number}{a non negative numeric value of the he minimum number of the reported modification containing bins.
If the bins are filtered less than this number by the p values or effect sizes,
more sites will be reported by the order of the p value until it reaches this number; Default to be floor( nrow(SE_bins)*0.002 ).}
}
\value{
a list of the index for the significant modified peaks (IP > input) and control peaks (peaks other than modification containing peaks).
}
\description{
\code{GLM_inference} conduct inference on log2 fold changes of IP over input using the negative binomial model in DESeq.
}
