% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/glmM.R
\docType{methods}
\name{glmM}
\alias{glmM}
\title{Quantification of RNA modification log2 fold change values based on the generalized linear models of negative binomial distribution.}
\usage{
glmM(sep, glm_type = c("auto", "poisson", "NB", "DESeq2"),
  LFC_shrinkage = c("apeglm", "Gaussian", "ashr"), ...)

\S4method{glmM}{SummarizedExomePeak}(sep, glm_type = c("auto", "poisson",
  "NB", "DESeq2"), LFC_shrinkage = c("apeglm", "Gaussian", "ashr",
  "none"), ...)
}
\arguments{
\item{sep}{is a summarizedExomePeak object.}

\item{glm_type}{a character, that specify the type of generalized linear model used in peak calling; The argument can be one of the "auto", "poisson", "NB", and "DESeq2". Default to be "auto".

Under the default setting, the DESeq2 GLM of NB is used on experiments with at least 3 biological replicates for both IP and input samples of the treatment and control.
The poisson GLM will be applied in the other cases.}

\item{LFC_shrinkage}{a character indicating the method for emperical bayes shrinkage on the log2 fold change estimates, can be one in "apeglm", "Gaussian", "ashr", or "none".
Please check \code{\link{lfcShrink}} for more information.}

\item{...}{Optional arguments passed to \code{\link{DESeq}}}
}
\description{
This function conduct a second round of RNA modification level quantification with generalized linear model estimates of the
neative binomial distribution. The RNA modification level is quantified as the log2 IP to input fold change.

By default, the modification level estimates will undergoes emperical Bayes shrinkage with a Couchey prior, which is defined in the package \link{apeglm}.
}
