% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/glmMeth.R
\docType{methods}
\name{glmMeth}
\alias{glmMeth}
\alias{glmMeth}
\title{Quantification of RNA methylation beta values based on the generalized linear models of negative binomial distribution.}
\usage{
glmMeth(sep, shrinkage_method = c("apeglm", "normal", "ashr"), ...)

\S4method{glmMeth}{SummarizedExomePeak}(sep, shrinkage_method = c("apeglm",
  "normal", "ashr"), ...)
}
\arguments{
\item{sep}{is a summarizedExomePeak object.}

\item{shrinkage_method}{a character indicating the method for emperical bayes shrinkage, can be one in "apeglm","normal", and "ashr".
Please check \code{\link{lfcShrink}} for more information.}

\item{...}{inherited arguments from \code{\link{DESeq}}}
}
\description{
This function conduct a second round of RNA methylation level quantification using generalized linear model estimates of the
neative binomial distribution. The RNA methyltion level is quantified as the log2 beta-value (log2 IP to input ratio).

By default, the methylation level estimates will undergoes emperical Bayes shrinkage using a Couchey prior, which is defined in the package \link{apeglm}.
}
