% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_plot.R
\name{guitar_plot}
\alias{guitar_plot}
\title{Plot the transcript distribution of the peaks}
\usage{
guitar_plot(sep, txdb = NULL, save_pdf_prefix = NULL,
  include_control_regions = TRUE, guitar_coordinate = NULL)
}
\arguments{
\item{sep}{a SummarizedExomePeaks object.}

\item{txdb}{a txdb object containing the transcript annotation.}

\item{save_pdf_prefix}{if provided, a pdf file with the given name will be saved under the current directory.}

\item{include_control_regions}{a logical indicating whether to include the control regions or not.}

\item{guitar_coordinate}{optional, the guitar coordinate of the transcript annotation.}
}
\value{
a ggplot object
}
\description{
This function plot the transcript topology of RNA modification based on the provided SummarizedExomePeaks object.
}
\details{
If the SummarizedExomePeaks object contains quantification results for methylation, the significantly methylated peaks
with IP to input log2FC > 0 and DESeq2 Wald test padj < 0.05 will be plotted.

If the SummarizedExomePeaks object contains quantification results for differential methylation, both the hyper methylation
and hypo methylation peaks with DESeq2 Wald test p values < 0.05 will be plotted.
}
