% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{LibraryType-methods}
\alias{LibraryType-methods}
\alias{LibraryType,MeripBamFileList-method}
\alias{LibraryType}
\title{Method LibraryType}
\usage{
LibraryType(x)

\S4method{LibraryType}{MeripBamFileList}(x)
}
\arguments{
\item{x}{a \code{MeripBamFileList} object.}
}
\description{
Method LibraryType

Accessor to the slot \code{LibraryType} in class \code{MeripBamFileList}.
}
