% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{LibraryType}
\alias{LibraryType}
\title{Accessor to the slot \code{LibraryType} in class \code{MeripBamFileList}.}
\usage{
LibraryType(x)

\S4method{LibraryType}{MeripBamFileList}(x)
}
\arguments{
\item{x}{a \code{MeripBamFileList} object.}
}
\value{
a logical vector.
}
\description{
Accessor to the slot \code{LibraryType} in class \code{MeripBamFileList}.
}
