% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_bg.R
\name{mclust_bg}
\alias{mclust_bg}
\title{Find background of merip-seq signal with model based clustering.}
\usage{
mclust_bg(se_peak_counts, alpha = 1)
}
\arguments{
\item{se_peak_counts}{A SummarizedExperiment object.}

\item{alpha}{The small offset added when calculating the M value.}
}
\value{
A dummy variable of the background index.
}
\description{
The function generate background index on the rowData of the summarizedExperiment.
}
\details{
This function will do the following jobs:

1. Filter the rows (modification sites) by average count.

2. Fit multivariate Gaussian mixture model with 2 mixing component to differentiate background and biological signal.

- depend on whether the bsgenome is provided, the values GC content will be used as one of the dimention of the clustering.

3. Classify the background and signal using bayes classifier by the learned model.

4. Return the index for the bins that are classified into background.
}
