% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/normalizeGC.R
\docType{methods}
\name{normalizeGC}
\alias{normalizeGC}
\title{Estimate Normalization Factors for GC content Bias Correction.}
\usage{
normalizeGC(sep, bsgenome = "hg19", txdb = "hg19", gff_dir = NULL,
  fragment_length = 100, binding_length = 25,
  feature = c("background", "all"), qtnorm = FALSE,
  effective_GC = FALSE)

\S4method{normalizeGC}{SummarizedExomePeak}(sep, bsgenome = NULL,
  txdb = NULL, gff_dir = NULL, fragment_length = 100,
  binding_length = 25, feature = c("all", "background"),
  qtnorm = TRUE, effective_GC = FALSE)
}
\arguments{
\item{sep}{a \code{\link{summarizedExomePeak}} object returned by \code{\link{exomePeak2}} or \code{\link{exomePeakCalling}}.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome reference,
If the \code{BSgenome} object is not available, it could be a \code{character} string of the UCSC genome name which is acceptable by \code{\link{getBSgenome}}, example: \code{"hg19"}.}

\item{txdb}{a \code{\link{TxDb}} object for the transcript annotation,
If the \code{TxDb} object is not available, it could be a \code{character} string of the UCSC genome name which is acceptable by \code{\link{makeTxDbFromUCSC}}, example: \code{"hg19"}.}

\item{fragment_length}{a positive integer number for the expected fragment length in nucleotides; default \code{= 100}.}

\item{binding_length}{a positive integer number for the expected binding length of the anti-modification antibody in IP samples; default \code{= 25}.}

\item{feature}{a \code{character} specifies the region used in the GC effect estimation, can be one in \code{c("all", "background")}; default \code{"all"}.

\describe{

 \item{\strong{\code{all}}}{
 The GC content and the parameters for quantile normalization will be estimated on all regions.
 }

 \item{\strong{\code{background}}}{
 The GC content linear effect and the parameters for quantile normalization will be estimated on the background regions.
 }

}}

\item{qtnorm}{a \code{logical} of whether to perform subset quantile normalization after the GC content linear effect correction； default \code{= TRUE}.

If \code{qtnorm = TRUE}, subset quantile normalization will be applied within the IP and input samples seperately to account for the inherent differences between the marginal distributions of IP and input samples.}

\item{effective_gc}{a \code{logical} of whether to calculate the effective GC content weighted by the fragment alignment probabilities; default \code{= FALSE}.}
}
\value{
a \code{summarizedExomePeak} object with the updated slot \code{GCsizeFactors}.
}
\description{
\code{normalizeGC} estimates the feature specific size factors in order to reduce the technical variation during modification peak statistics quantification.
}
\details{
PCR amplication bias related to GC content is a major source of technical variation in RNA-seq.
The GC content biases are usually correlated within the same laboratory environment, and this will result in the batch effect between different studies.

The GC content normalization can result in an improvement of peak accuracy for most published m6A-seq data,
and it is particullarly recommended if you want to compare the quantifications on methylation levels between different laboratory conditions.
}
\examples{

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)

aln <- scanMeripBAM(
bam_ip = c("IP_rep1.bam",
          "IP_rep2.bam",
          "IP_rep3.bam"),
bam_input = c("input_rep1.bam",
             "input_rep2.bam",
             "input_rep3.bam"),
paired_end = TRUE
)

sep <- exomePeakCalling(merip_bams = aln,
                       txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
                       bsgenome = Hsapiens)

sep <- normalizeGC(sep)

}
\seealso{
\code{\link{estimateSeqDepth}}
}
