% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{Parameter-methods}
\alias{Parameter-methods}
\alias{Parameter,MeripBamFileList-method}
\alias{Parameter}
\title{Method Parameter}
\usage{
Parameter(x)

\S4method{Parameter}{MeripBamFileList}(x)
}
\arguments{
\item{x}{a \code{MeripBamFileList} object.}
}
\description{
Method Parameter

Accessor to the slot \code{Parameter} in class \code{MeripBamFileList}.
}
