% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{Parameter}
\alias{Parameter}
\title{Accessor to the slot \code{Parameter} in class \code{MeripBamFileList}.}
\usage{
Parameter(x)

\S4method{Parameter}{MeripBamFileList}(x)
}
\arguments{
\item{x}{a \code{MeripBamFileList} object.}
}
\value{
a \code{ScanBamParam} object.
}
\description{
Accessor to the slot \code{Parameter} in class \code{MeripBamFileList}.
}
