% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotBetaGC.R
\docType{methods}
\name{plotBetaGC}
\alias{plotBetaGC}
\alias{plotBetaGC}
\title{Plot the relationship between the glm estimate and GC content.}
\usage{
plotBetaGC(sep, bsgenome, save_pdf_prefix = NULL, fragment_length = 100)

\S4method{plotBetaGC}{SummarizedExomePeak}(sep, bsgenome,
  save_pdf_prefix = NULL, fragment_length = 100)
}
\arguments{
\item{sep}{a \code{summarizedExomePeak} object.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence.}

\item{save_pdf_prefix}{a character, if provided, a pdf file with the given name will be saved under the current directory.}

\item{fragment_length}{the expected fragment length of the sequencing library.
The widths of the row features for quantifying GC content will be re-sized into the fragment length if it falls bellow it.}
}
\value{
a ggplot object.
}
\description{
This function plot the scatter plot between GC content and the methylation / differential methylation levels of the modification sites.
}
\details{
By default, this function will generate the countour of the scatter plot, and a linear regression line indicating the trend between
GC content and log2 fold change or log2 odds ratio returned by DESeq2.
The significant changed methylation sites will be grouped and lebeled in different colours.
}
