% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotLfcGC.R
\docType{methods}
\name{plotLfcGC-methods}
\alias{plotLfcGC-methods}
\alias{plotLfcGC,SummarizedExomePeak-method}
\alias{plotLfcGC}
\title{Method plotLfcGC}
\usage{
plotLfcGC(sep, bsgenome = NULL, txdb = NULL, save_pdf_prefix = NULL,
  fragment_length = 100, binding_length = 25, effective_GC = FALSE,
  save_dir = ".")

\S4method{plotLfcGC}{SummarizedExomePeak}(sep, bsgenome = NULL,
  txdb = NULL, save_pdf_prefix = NULL, fragment_length = 100,
  binding_length = 25, effective_GC = FALSE, save_dir = ".")
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeak}} object.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence, it could be the name of the reference genome recognized by \code{\link{getBSgenome}}.}

\item{txdb}{a \code{\link{TxDb}} object for the transcript annotation, it could be the name of the reference genome recognized by \code{\link{makeTxDbFromUCSC}}.}

\item{save_pdf_prefix}{a \code{character}, if provided, a pdf file with the given name will be saved under the current directory; Default \code{= NULL}.}

\item{fragment_length}{a \code{numeric} value for the expected fragment length in the RNA-seq library; Default \code{= 100}.}

\item{binding_length}{a \code{numeric} value for the expected antibody binding length in IP samples; Default \code{= 25}.}

\item{effective_GC}{a \code{logical} value of whether to calculate the weighted GC content by the probability of reads alignment; default \code{= FALSE}.}

\item{save_dir}{a \code{character} for the directory to save the plot; default ".".}
}
\value{
a \code{ggplot} object.
}
\description{
\code{plotLfcGC} plot the scatter plot between GC content and the (differential) modification LFCs.
}
\details{
By default, this function will generate a scatter plot between GC content and the log2FC value.
The significant modification sites will be lebeled in different colours.
}
\examples{

\dontrun{
sep #a SummarizedExomePeak object
plotLfcGC(sep)
}

}
