% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSizeFactors.R
\name{plotSizeFactors-methods}
\alias{plotSizeFactors-methods}
\alias{plotSizeFactors,SummarizedExomePeak-method}
\alias{plotSizeFactors}
\title{Method plotSizeFactors}
\usage{
plotSizeFactors(sep)

\S4method{plotSizeFactors}{SummarizedExomePeak}(sep)
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeak}} object.}
}
\value{
A ggplot.
}
\description{
Method plotSizeFactors

plot the size factors using different strategies.
}
\examples{

### Load the example SummarizedExomPeak object
f1 = system.file("extdata", "sep_ex_mod.rds", package="exomePeak2")

sep <- readRDS(f1)

### Visualize the size factors estimated using different regions
plotSizeFactors(sep)

}
