% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSizeFactors.R
\docType{methods}
\name{plotSizeFactors}
\alias{plotSizeFactors}
\title{plot the size factors using different strategies.}
\usage{
plotSizeFactors(sep)

\S4method{plotSizeFactors}{SummarizedExomePeak}(sep)
}
\description{
plot the size factors using different strategies.
}
