% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{RandomPrimer}
\alias{RandomPrimer}
\alias{RandomPrimer}
\title{Accessor to the slot \code{RandomPrimer} for the object \code{MeripBamFileList}.}
\usage{
RandomPrimer(x)

\S4method{RandomPrimer}{MeripBamFileList}(x)
}
\arguments{
\item{x}{a \code{MeripBamFileList} object.}
}
\value{
a logical vector.
}
