% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_five_POS.R
\name{reads_five_POS}
\alias{reads_five_POS}
\title{extract reads five prime POS.}
\usage{
reads_five_POS(reads, width = 1, fix = "start", ...)
}
\arguments{
\item{reads}{a GAlignmentList object.}

\item{width}{the width of the five prime POS.}

\item{fix}{the end of the POS, start means 5', end means 3'.}

\item{...}{arguments path to function resize.}
}
\value{
the resized reads GRanges for 5' POS positions.
}
\description{
extract reads five prime POS.
}
