% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_peaks.R
\name{reduce_peaks}
\alias{reduce_peaks}
\title{A function to reduce ranges between elements within a GRangesList.}
\usage{
reduce_peaks(peaks_grl, txdb)
}
\arguments{
\item{peaks_grl}{The GRangesList of the peaks region to be reduced.}

\item{txdb}{The txdb object used during the generation of the peaks.}
}
\value{
GRangesList object that is reduced between its inner elements.
The metadata collumn is preserved, however the names is not tracked anymore.
}
\description{
\code{reduce_inter_grl} redivide and reduce the overlapping GRangesList element.
}
\keyword{internal}
