% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_introns.R
\name{remove_introns}
\alias{remove_introns}
\title{Removing introns from a provided GRanges object.}
\usage{
remove_introns(gr_bins, grl_exbg)
}
\arguments{
\item{gr_bins}{A \code{GRanges} object of exomePeak bins before the intron removal.}

\item{grl_exbg}{A \code{GRangesList} object that define the exon regions of each genes.}
}
\value{
A \code{GRangesList} object with the same length of \code{gr_bins}, each list element corresponds to the original GRanges after the removal of introns.
}
\description{
Removing introns from a provided GRanges object.
}
\keyword{internal}
