% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_path.R
\name{save_path}
\alias{save_path}
\title{Set the path to save files across different systems.}
\usage{
save_path(save_dir, save_file_name, file_end)
}
\arguments{
\item{save_dir}{directory to save the file.}

\item{save_file_name}{name of the file being saved.}

\item{file_end}{end of the file being saved.}
}
