% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Export the methylation / differential methylation sites}
\usage{
save_results(SEP, format = c("txt", "BED", "RDS"),
  file_name = "exomepeaks_result", cut_off_pvalue = NULL,
  cut_off_padj = 0.05, cut_off_log2FC = 0, min_num_of_positive = 2000,
  expected_direction = "both", inhibit_filter = FALSE)
}
\arguments{
\item{SEP}{A SummarizedExomePeak object.}

\item{format}{A character string of one of the c("txt", "BED", "RDS"), indicating the exported format.

- The choice txt will save a tab separated values (tsv) file with the ranges and test information.

- The choice BED will save a BEDGraph file with the score column being the adjusted p value.

- The choice RDS will save a Rdata of the SummarizedExperiment object which includes a comprehensive summary of count and design information (recommended).}

\item{file_name}{The name of the file saved; Default "exomepeaks_result".}

\item{cut_off_pvalue}{A number between 0 and 1 indicate the p value cutoff in the exported result; Default NULL.}

\item{cut_off_padj}{A number between 0 and 1 indicate the adjusted p value cutoff in the exported result; Default 0.05.}

\item{cut_off_log2FC}{A non negative number indicating the log2 fold change cutoff of the exported result,
only sites with log2 IP/input fold change bigger than this value are kept; Default 0.

For differential methylation analysis, the absolute value of the log2 Odds ratio will be filtered by \code{cut_off_log2FC}.}

\item{min_num_of_positive}{The minimum number of reported sites.
If the sites is filtered less than this number by its p values or effect sizes,
more sites will be reported by the order of the p value until it reaches this number.}

\item{expected_direction}{The expected differential methylation direction, could be "hyper", "hypo", or "both".
This argument is useful when the treated group involves the perturbation of a writer or eraser protein for the modification; Default "both".}

\item{inhibit_filter}{Remove all the filters upon on the result, this is desired for user provided modification annotation; Default FALSE.}
}
\value{
a file containing the modification site genomic location, gene ids, statistics, and effect sizes will be created under the current working directory.
}
