% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sep.R
\name{sort_sep}
\alias{sort_sep}
\title{a function to sort the summarizedExomePeak by its order on genome}
\usage{
sort_sep(sep)
}
\arguments{
\item{sep}{a summarizedExomePeak object.}
}
\value{
the summarizedExomePeak with a given order.
}
\description{
a function to sort the summarizedExomePeak by its order on genome
}
\keyword{internal}
