% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummarizedExomePeak-class}
\alias{SummarizedExomePeak-class}
\title{SummarizedExomePeak}
\description{
An object that summarizes the methylation sites information,
size factors for technical variabilities, and the result of the DESeq2 inference.
}
