% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gene_bins.R
\name{visualize_gene_bins}
\alias{visualize_gene_bins}
\title{A convenient function to help you quickly visualize the exome bins on a gene.}
\usage{
visualize_gene_bins(gene_id, gr_bins, grl_gene)
}
\arguments{
\item{gene_id}{The id of the gene.}

\item{gr_bins}{A \code{GRanges} object of exomePeak bins.}

\item{grl_exbg}{A \code{GRangesList} object that define the exon regions of each genes.}
}
\value{
A Gvis diagram.
}
