% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTxDb.R
\name{convertTxDb}
\alias{convertTxDb}
\title{Convert the txdb object into the full transcript and whole genome types}
\usage{
convertTxDb(txdb, type = c("full_tx", "whole_genome"))
}
\arguments{
\item{txdb}{a \code{TxDb} object containing the regular transcript annotation.}

\item{type}{the type of \code{TxDb} object of the output, can be one in c("full_tx","whole_genome").}
}
\value{
a TxDb object that will change the exon region into the full transcript and the whole genome regions.
}
\description{
This function can convert the txdb object into full transcript and whole genome types.
}
\details{

}
