% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disj_background.R
\name{disj_background}
\alias{disj_background}
\title{Find the background of the user provided mod_gration.}
\usage{
disj_background(mod_gr, txdb, cut_off_num = 30, background_bins = NULL,
  background_types = c("mclust", "m6Aseq_prior", "manual", "all"),
  control_width = 50, rename_mod = FALSE)
}
\arguments{
\item{mod_gr}{A \code{GRanges} object of user provided mod_gration (names are neccessary for the index of the splitting).}

\item{txdb}{A \code{TxDb} object that define the transcript mod_gration.}

\item{cut_off_num}{A non-negative integer indicate the leat total number of the disjoint exons used as the background; Default 30.}

\item{rename_mod}{Whether to rename the returned modification sites, default = FALSE.}

\item{m6Aseq_background}{A logical value, TRUE if the region of 5'UTR and long exons of the transcripts should be dropped in control region; Default TRUE.}

\item{distance_5p}{A numeric value of the length of the transcript starting region; default 200.}
}
\value{
A \code{GRangesList} object.
The first portion is the exons regions that is not overlapped with \code{annoation}.

If the resulting ranges have less number and width compared with what defined in \code{cut_off_num},
the exon regions of txdb will be returned as the background.

The second portion is the restructed user provided mod_gration with gene id mod_grated.
}
