% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estimateSeqDepth.R
\docType{methods}
\name{estimateSeqDepth-methods}
\alias{estimateSeqDepth-methods}
\alias{estimateSeqDepth,SummarizedExomePeak-method}
\alias{estimateSeqDepth}
\title{Method estimateSeqDepth}
\usage{
estimateSeqDepth(sep, from = c("Background", "Modification", "All"), ...)

\S4method{estimateSeqDepth}{SummarizedExomePeak}(sep,
  from = c("Background", "Modification", "All"), ...)
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeak}} object.}

\item{from}{a \code{character} specify the subset of features for sequencing depth estimation, can be one of \code{c("Background", "Modification", "All")}.

\describe{
 \item{\strong{\code{Background}}}{
 The sequencing depths are estimated from the background regions. This could make the IP/input LFC estimates become a rescaled version of the real modification proportion.
 }

 \item{\strong{\code{Modification}}}{
 The sequencing depths are estimated from the modification peaks/sites.
 }

 \item{\strong{\code{All}}}{
 The sequencing depths are estimated from both the background and modification regions.
 }
}

Under default setting, the sequencing depth factors are estimated from the background Background regions.}

\item{...}{inherited from \code{\link{estimateSizeFactorsForMatrix}}.}
}
\value{
This function will return a \code{\link{SummarizedExomePeak}} object containing newly estimated sequencing depth size factors.
}
\description{
\code{estimateSeqDepth} estimate sequencing depth size factors for each MeRIP-seq samples.
Under default setting, the sequencing depth are estimated by the robust estimator defined in package DESeq2.
i.e. the median of the ratios to the row geometric means.
}
\details{
The function takes the input of a \code{\link{SummarizedExomePeak}} object,
and it estimates the sequencing depth size factors by the columns of its \link{assay}.
}
\examples{

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)

aln <- scanMeripBAM(
bam_ip = c("IP_rep1.bam",
           "IP_rep2.bam",
           "IP_rep3.bam"),
bam_input = c("input_rep1.bam",
              "input_rep2.bam",
              "input_rep3.bam"),
paired_end = TRUE
)

sep <- exomePeakCalling(merip_bams = aln,
                       txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
                       bsgenome = Hsapiens)

sep <- estimateSeqDepth(sep)

}
\seealso{
\code{\link{normalizeGC}}
}
