% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exomePeakCalling.R
\docType{methods}
\name{exomePeakCalling}
\alias{exomePeakCalling}
\title{Perform Peak Calling on MeRIP-seq Dataset.}
\usage{
exomePeakCalling(merip_bams = NULL, txdb = NULL, bsgenome = NULL,
  mod_annot = NULL, glm_type = c("DESeq2", "NB", "Poisson"),
  background = c("Gaussian_mixture", "m6Aseq_prior", "manual", "all"),
  manual_background = NULL, gff_dir = NULL, fragment_length = 100,
  binding_length = 25, step_length = binding_length,
  pc_count_cutoff = 5, bg_count_cutoff = 50, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0,
  peak_width = fragment_length/2, parallel = FALSE, bp_param = NULL)

\S4method{exomePeakCalling}{MeripBamFileList}(merip_bams = NULL,
  txdb = NULL, bsgenome = NULL, mod_annot = NULL,
  glm_type = c("DESeq2", "NB", "Poisson"),
  background = c("Gaussian_mixture", "m6Aseq_prior", "manual", "all"),
  manual_background = NULL, gff_dir = NULL, fragment_length = 100,
  binding_length = 25, step_length = binding_length,
  pc_count_cutoff = 5, bg_count_cutoff = 50, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0,
  peak_width = fragment_length/2, parallel = FALSE, bp_param = NULL)
}
\arguments{
\item{merip_bams}{a \code{MeripBamFileList} object returned by \code{\link{scanMeripBAM}}.}

\item{txdb}{a \code{\link{TxDb}} object for the transcript annotation,
If the \code{TxDb} object is not available, it could be a \code{character} string of the UCSC genome name which is acceptable by \code{\link{makeTxDbFromUCSC}}, example: \code{"hg19"}.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence information,
If the \code{BSgenome} object is not available, it could be a \code{character} string of the UCSC genome name which is acceptable by \code{\link{getBSgenome}}, example: \code{"hg19"}.}

\item{mod_annot}{a \code{\link{GRanges}} object for user provided single based RNA modification annotation.

If user provides the single based RNA modification annotation, this function will perform reads count on the provided annotation flanked by length \code{= floor(fragment_length - binding_length/2)}.}

\item{glm_type}{a \code{character} speciefies the type of Generalized Linear Model (GLM) fitted for the purpose of statistical inference during peak calling, which can be one of the \code{c("DESeq2", "NB", "Poisson")}.

\describe{
\item{\strong{\code{DESeq2}}}{Fit the GLM defined in function \code{\link{DESeq}}, which is the NB GLM with regulated estimation of the overdispersion parameters.}

\item{\strong{\code{NB}}}{Fit the Negative Binomial (NB) GLM.}

\item{\strong{\code{Poisson}}}{Fit the Poisson GLM.}
}

By default, the DESeq2 GLMs are fitted on the data set with > 1 biological replicates for both the IP and input samples, the Poisson GLM will be fitted otherwise.}

\item{background}{a \code{character} specifies the method for the background finding, i.e. to identify the windows without modification signal. It could be one of \code{c("Gaussian_mixture", "m6Aseq_prior", "manual", "all")};  default \code{= "Gaussian_mixture"}.

In order to accurately account for the technical variations, it is often neccessary to estimate the sequencing depth and GC content linear effects on windows without modification signals.

The following methods are supported in \code{ExomePeak2} to differentiate the no modification background windows from the modification containig windows.

\describe{
 \item{\strong{\code{Gaussian_mixture}}}{The background is identified by Multivariate Gaussian Mixture Model (MGMM) with 2 mixing components on the vectors containing methylation level estimates and GC content, the background regions are predicted by the Bayes Classifier on the learned GMM.}

 \item{\strong{\code{m6Aseq_prior}}}{The background is identified by the prior knowledge of m6A topology, the windows that are not overlapped with long exons (exon length >= 400bp) and 5'UTR are treated as the background windows.

 This type of background should not be used if the MeRIP-seq data is not targetting on m6A methylation.

 }

 \item{\strong{\code{manual}}}{The background regions are defined by the user manually at the argument \code{manual_background}.}

 \item{\strong{\code{all}}}{Use all windows as the background. This is equivalent to not differentiating background and signal.
 It can lead to biases on the estimation of the technical factors.
 }
}}

\item{manual_background}{optional, a \code{\link{GRanges}} object for the user provided unmodified background.}

\item{gff_dir}{optional, a \code{character} which specifies the directory toward a gene annotation GFF/GTF file, it is applied when the \code{TxDb} object is not available, default \code{= NULL}.}

\item{fragment_length}{a positive integer number for the expected fragment length in nucleotides; default \code{= 100}.}

\item{binding_length}{a positive integer number for the expected binding length of the anti-modification antibody in IP samples; default \code{= 25}.}

\item{step_length}{a positive integer number for the shift distances of the sliding window; default \code{= binding_length}.}

\item{pc_count_cutoff}{a \code{numeric} value for the cutoff on average window's reads count in peak calling; default \code{= 5}.}

\item{bg_count_cutoff}{a \code{numeric} value for the cutoff on average window's reads count in background identification; default \code{= 50}.}

\item{p_cutoff}{a \code{numeric} value for the cutoff on p values in peak calling; default \code{= NULL}.}

\item{p_adj_cutoff}{a \code{numeric} value for the cutoff on Benjamini Hochberg adjusted p values in peak calling; default \code{= 0.05}.}

\item{logFC_cutoff}{a \code{numeric} value for the cutoff on log2 IP over input fold changes in peak calling; default \code{= 0}.}

\item{peak_width}{a \code{numeric} value for the minimum width of the merged peaks; default \code{= fragment_length} .}

\item{parallel}{a \code{logical} value indicating whether to use parallel computation, typlically it requires more than 16GB of RAM if \code{parallel = TRUE}; default \code{= FALSE}.}

\item{bp_param}{optional, a \code{\link{BiocParallelParam}} object that stores the configuration parameters for the parallel execution.}
}
\value{
a \code{\link{SummarizedExomePeak}} object.
}
\description{
\code{exomePeakCalling} call peaks of RNA modification from a MeRIP-seq data set.
}
\details{
\code{exomePeakCalling} perform peak calling from the MeRIP-seq BAM files on exon regions defined by the user provided transcript annotations.
If the \code{\link{BSgenome}} object is provided, the peak calling will be conducted with the GC content bias correction.

Under the default setting, for each window, exomePeak2 will fit a GLM of Negative Binomial (NB) with regulated estimation of the overdispersion parameters developed in \code{\link{DESeq}}.
Wald tests with H0 of IP/input Log2 Fold Change (LFC) <= 0 are performed on each of the sliding windows.
The significantly modified peaks are selected using the cutoff of fdr < 0.05 calculated by the Benjamini Hochberg approach.
}
\examples{

# Load packages for the genome sequence and transcript annotation

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)

# Peak Calling

exomePeakCalling(
  merip_bams = scanMeripBAM(
  bam_ip = c("IP_rep1.bam",
             "IP_rep2.bam",
             "IP_rep3.bam"),
  bam_input = c("input_rep1.bam",
                "input_rep2.bam",
                "input_rep3.bam"),
  paired_end = TRUE),
  txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
  bsgenome = Hsapiens)

# Analysis with single based modification annotation

annot_dir <- system.file("extdata", "m6A_hg19_annot.rds", package = "exomePeak2")

m6A_hg19_gr <- readRDS(annot_dir)

exomePeakCalling(
  merip_bams = scanMeripBAM(
  bam_ip = c("IP_rep1.bam",
             "IP_rep2.bam",
             "IP_rep3.bam"),
  bam_input = c("input_rep1.bam",
                "input_rep2.bam",
                "input_rep3.bam"),
 paired_end = TRUE),
 txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
 bsgenome = Hsapiens,
 mod_annot = m6A_hg19_gr)


}
\seealso{
\code{\link{exomePeak2}}, \code{\link{glmM}}, \code{\link{glmDM}}, \code{\link{normalizeGC}}, \code{\link{exportResults}}
}
