% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exomePeak2.R,
%   R/exomePeakCalling.R
\docType{methods}
\name{exomePeakCalling}
\alias{exomePeakCalling}
\alias{exomePeakCalling}
\alias{exomePeak2}
\alias{exomePeakCalling}
\title{Exome peak calling and quatification of MeRIP-seq datasets.}
\usage{
exomePeakCalling(merip_bams = NULL, txdb = NULL, bsgenome = NULL,
  glm_type = c("DESeq2", "NB", "poisson"), background = c("mclust",
  "m6Aseq_prior", "manual", "all"), manual_background = NULL,
  gene_annot = NULL, mod_annot = NULL, fragment_length = 100,
  binding_length = 25, step_length = binding_length, pc_count_cutoff = 5,
  gc_count_cutoff = 50, p_cutoff = NULL, p_adj_cutoff = 0.05,
  logFC_cutoff = 0, peak_width = fragment_length/2, parallel = FALSE,
  bp_param = NULL)

exomePeak2(bam_ip = NULL, bam_input = NULL, bam_treated_ip = NULL,
  bam_treated_input = NULL, txdb = NULL, bsgenome = NULL,
  genome_assembly = NA, gene_annot = NULL, mod_annot = NULL,
  paired_end = FALSE, random_primer = TRUE, fragment_length = 100,
  binding_length = 25, step_length = binding_length,
  peak_width = fragment_length/2, pc_count_cutoff = 5,
  gc_count_cutoff = 50, p_cutoff = NULL, p_adj_cutoff = 0.05,
  logFC_cutoff = 0, parallel = FALSE, background = c("mclust",
  "m6Aseq_prior", "manual", "all"), glm_type = c("DESeq2", "poisson", "NB"),
  LFC_shrinkage = c("apeglm", "ashr", "Gaussian", "none"),
  export_results = TRUE, export_format = c("tsv", "BED", "RDS"),
  table_style = c("bed", "granges"), save_plot_GC = TRUE,
  save_plot_analysis = FALSE, save_plot_name = "ep2",
  save_dir = "exomePeak2_output")

\S4method{exomePeakCalling}{MeripBamFileList}(merip_bams = NULL,
  txdb = NULL, bsgenome = NULL, glm_type = c("DESeq2", "NB", "poisson"),
  background = c("mclust", "m6Aseq_prior", "manual", "all"),
  manual_background = NULL, gene_annot = NULL, mod_annot = NULL,
  fragment_length = 100, binding_length = 25,
  step_length = binding_length, pc_count_cutoff = 5, gc_count_cutoff = 50,
  p_cutoff = NULL, p_adj_cutoff = 0.05, logFC_cutoff = 0,
  peak_width = fragment_length/2, parallel = FALSE, bp_param = NULL)
}
\arguments{
\item{merip_bams}{a \code{MeripBamFileList} object.}

\item{txdb}{a txdb object, it could be a single character string that can be recognized by \code{\link{makeTxDbFromUCSC}}; Default "hg19".}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence, alternatively it could be the name of the reference genome recognized by \code{\link{getBSgenom}}.}

\item{glm_type}{a character, which can be one of the "auto", "poisson", "NB", and "DESeq2". This argument specify the type of generalized linear model used in peak calling; Default to be "auto".}

\item{background}{a \code{GRanges} or \code{GRangesList} object for user provided background control regions on the genome.}

\item{manual_background}{a \code{GRanges} object for user provided RNA modification background.}

\item{mod_annot}{a \code{GRanges} object for user provided single based RNA modification annotation. If provided, the peak calling step will be skipped.}

\item{fragment_length}{a positive integer of the expected fragment length in bp; default 100.}

\item{binding_length}{a positive integer of the antibody binding length in IP samples; default 25.}

\item{step_length}{a positive integer of the shift size of the sliding window; default is the binding length.}

\item{pc_count_cutoff}{a non negative integer value of the average reads count per window used in peak calling; default 5.}

\item{gc_count_cutoff}{a non negative integer value of the minimum average reads count per window used in GC effect estimation; default 50.}

\item{p_cutoff}{a value of the p value cut-off used in peak calling; default NULL.}

\item{p_adj_cutoff}{a value of the adjusted p value cutoff used in DESeq inference; default 0.05.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{peak_width}{positive integer of the minimum width for the merged peaks; default \code{fragment_length} .}

\item{parallel}{a logical indicating whether to use parallel computation, consider this if your computer has more than 16GB RAM.}

\item{save_plot_GC}{a logical indicating whether to generate and save the plots of GC content bias assessment; default TRUE.}

\item{save_plot_analysis}{a logical indicating whether to generate and save the plots of genomic analysis on modification sites; default FALSE.}

\item{save_plot_name}{a character for the name of the plots being saved; Default "ep2".}

\item{save_dir}{a character for the name of the directory being saved; Default "exomePeak2_output".
Reads count will be performed using the provided annotation flanked by length of floor(fragment_length - binding_length/2).

The background regions used in this senario will be the disjoint exon regions of the flanked provided sites.}

\item{width_cutoff}{a positive integer of the minimum width for the merged peaks; default \code{fragment_length} .}

\item{merip_bams}{a \code{MeripBamFileList} object.}

\item{txdb}{a \code{TxDb} object, it can also be a single character string such as "hg19" which is processed by \code{\link{makeTxDbFromUCSC}}.}

\item{gene_anno}{a string, which specifies a gene annotation GFF/GTF file if available, default: NA.}

\item{fragment_length}{a positive integer of the expected fragment length in bp; default 100.}

\item{binding_length}{a positive integer of the antibody binding length in IP samples; default 25.}

\item{step_length}{a positive integer of the shift size of the sliding window; default is the binding length.}

\item{glm_type}{a character, which can be one of the "DESeq2", "poisson", "NB". This argument specify the type of generalized linear model used in peak calling; Default to be "DESeq2".

Under the default setting, the DESeq2 method is implemented on experiments with > 1 biological replicates for both IP and input samples.
The poisson GLM will be implemented otherwise.}

\item{pc_count_cutoff}{a non negative integer value of the minimum average reads count per window used in peak calling; default 5.}

\item{p_cutoff}{a value of the p value cut-off used in peak calling; default NULL.}

\item{p_adj_cutoff}{a value of the adjusted p value cutoff used in DESeq inference; default 0.05.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{parallel}{a logical indicating whether to use parallel computation, consider this if your computer has more than 16GB RAM.}

\item{mod_annot}{a \code{GRanges} object for user provided single based RNA modification annotation. If provided, the peak calling step will be skipped.
Reads count will be performed using the provided annotation flanked by length of floor(fragment_length - binding_length/2).}

\item{m6Aseq_background}{a logical of whether to use the topology knowledge of m6A-Seq to find appropriate background in GC effect estimation;
If TRUE, the GC effect will be estimated on bins that is not overlapping with long exons (exon length >= 400bp) and 5'UTR.
Also, the returned background control ranges returned will also exclude those regions;
It should not be select if you are analyzing MeRIP-seq data of other modification, such as hm5C; default TRUE.

The background regions used in this senario will be the disjoint exon regions of the flanked provided sites.}
}
\value{
This function will save the results of modification / differrential modification analysis under the current working directory.

This function will return a \code{SummarizedExomePeak} object storing the ranges and reads counts information of the called peaks.
}
\description{
This function perform a default exomePeak2 analysis with the following steps:

 1. scan MeRIP-seq BAM files: \code{\link{scanMeripBAM}}.

 2. call modification peaks on exons: \code{\link{exomePeakCalling}}.

 3. gc correction based on non-modification background regions: \code{\link{normalizeGC}}.

 4. inference and quantify modification / differential modification with Negative Binomial GLM:  \code{\link{glmM}}; \code{\link{glmDM}}

 For complete details on each step, please check the manual pages for the corresponding functions.

  call RNA modification peaks and quantify the RNA modification levels on the exon regions.
If the treatment samples are provided, differential modification analysis will be conducted to infer the dynamics of RNA modification upon treatment condiftions.

\code{exomePeakCalling} call RNA modification peaks on exome regions with statistical tests that account for the biological variabilities between samples.
}
\details{
\code{exomePeak2} call peaks on exonic regions while adjust for GC content biases.
A GFF file or a TxDB object should be provided for transcript annotation.
Additionally, a BSgenome object is required to perform the GC content adjustment on the modification peaks.

The function conduct exome level peak calling based on the read alignment results and the transcript annotations.
}
\examples{




}
\seealso{
\code{\link{exomePeakCalling}}

\code{\link{exomePeakCalling}}
}
