% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exomePeakCalling.R
\docType{methods}
\name{exomePeakCalling}
\alias{exomePeakCalling}
\alias{exomePeakCalling}
\title{Exome peak calling on MeRIP-seq datasets while considering the biological variabilities.}
\usage{
exomePeakCalling(merip_bams = NULL, genome = "hg19", gff_dir = NULL,
  txdb = NULL, window_size = 100, step_size = 10, count_cutoff = 10,
  p_cutoff = NULL, p_adj_cutoff = 0.05, logFC_cutoff = 0,
  drop_overlapped_genes = TRUE, parallel = FALSE, mod_annotation = NULL,
  background = NULL)

\S4method{exomePeakCalling}{MeripBamFileList}(merip_bams = NULL,
  genome = "hg19", gff_dir = NULL, txdb = NULL, window_size = 100,
  step_size = 10, count_cutoff = 10, p_cutoff = NULL,
  p_adj_cutoff = 0.05, logFC_cutoff = 0, drop_overlapped_genes = TRUE,
  parallel = FALSE, mod_annotation = NULL, background = NULL)
}
\arguments{
\item{merip_bams}{a \code{MeripBamFileList} object.}

\item{genome}{a single character string describing the genome assembly in UCSC, by default it is "hg19". (not implemented yet)}

\item{gff_dir}{the path of the GFF or GTF file directory for transcript annotation.}

\item{txdb}{a txdb object, if provided, it will cover the annotation files defined above.}

\item{window_size}{a positive integer value of the window or bin width in bp.}

\item{step_size}{a positive integer value of the step or shift width in bp.}

\item{count_cutoff}{a non negative integer value of the minimum total reads count per window used in peak calling, default to be 10.}

\item{p_cutoff}{a value of the p value cut-off used in peak calling, default to be 0.05.}

\item{p_adj_cutoff}{a value of the adjusted p value cutoff used in DESeq inference; if provided, the value of \code{p_cutoff} will be ignored.}

\item{logFC_cutoff}{a non negative numeric value of the log2 fold change (log2 IP/input) cutoff used in the inferene of peaks.}

\item{drop_overlapped_genes}{a logical indicating whether the bins on overlapping genes are dropped or not, default to be TRUE.}

\item{parallel}{a logical indicating whether to use parallel computation, consider this if your computer has more than 16GB RAM.}

\item{mod_annotation}{a \code{GRanges} or \code{GRangesList} object for user provided RNA modification annotation. If provided, the peak calling step will be skipped.
Reads counting will be performed using the provided annotation, annotations with size less than \code{window_size} are resized to \code{window_size} fixed at the center.

The background regions used in this senario will be the exon ranges other than the provided annotation.}

\item{background}{a \code{GRanges} or \code{GRangesList} object for user provided background control regions on the genome.}
}
\value{
This function will return a \code{SummarizedExomePeak} object storing the ranges and reads counts information of the called peaks.
}
\description{
\code{exomePeakCalling} call RNA methylation peaks on exome regions with statistical tests that account for the biological variabilities between samples.
}
\details{
The function conduct exome level peak calling based on the read alignment results and the transcript annotations.
}
\examples{


}
\seealso{
\code{\link{exomePeakCalling}}
}
