% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonPlot.R
\name{exonPlot}
\alias{exonPlot}
\title{plot the distribution for the length of the overlapped exons.}
\usage{
exonPlot(gfeatures, txdb, save_pdf_prefix = NULL, save_dir = ".")
}
\arguments{
\item{gfeatures}{a list of GRanges or GRangesList.}

\item{txdb}{a txdb object.}

\item{save_pdf_prefix}{provided to save a pdf file.}

\item{save_dir}{a character indicating the directory to save the plot; default ".".}
}
