% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exons_by_unique_gene.R
\name{exons_by_unique_gene}
\alias{exons_by_unique_gene}
\title{Extracting exons by the corresponding genes that are on the same chromosomes and strand.}
\usage{
exons_by_unique_gene(txdb, drop_overlapped_genes = TRUE)
}
\arguments{
\item{txdb}{A TXDB object.}

\item{drop_overlapped_genes}{A logical value, TRUE if the overlapped genes are deleted from the output.}
}
\value{
A \code{GRangesList} object, each element in it corresponds to GRanges of the reduced exons of an unique gene,
the name corresponds to the original gene with .integer indexed if they have exons on different strands and chromosomes.
}
