% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Allaccessors.R
\docType{methods}
\name{GCsizeFactors}
\alias{GCsizeFactors}
\title{Accessor to the slot \code{GCsizeFactors} in class \code{SummarizedExomePeak}.}
\usage{
GCsizeFactors(x)

\S4method{GCsizeFactors}{SummarizedExomePeak}(x)
}
\arguments{
\item{x}{A \code{SummarizedExomePeak} object.}
}
\value{
A matrix.
}
\description{
Accessor to the slot \code{GCsizeFactors} in class \code{SummarizedExomePeak}.
}
