% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/glmDM.R
\docType{methods}
\name{glmDM}
\alias{glmDM}
\title{Estimation and Inference on RNA Differential Modification Log2 Fold Changes with Generalized Linear Model.}
\usage{
glmDM(sep, glm_type = c("auto", "Poisson", "NB", "DESeq2"),
  LFC_shrinkage = c("apeglm", "ashr"), ...)

\S4method{glmDM}{SummarizedExomePeak}(sep, glm_type = c("auto",
  "Poisson", "NB", "DESeq2"), LFC_shrinkage = c("apeglm", "ashr",
  "none"), ...)
}
\arguments{
\item{sep}{a \code{\link{summarizedExomePeak}} object.}

\item{glm_type}{a \code{character} speciefies the type of Generalized Linear Model (GLM) fitted for the purpose of statistical inference during peak calling, which can be one of the \code{c("DESeq2", "NB", "Poisson")}.

\describe{
\item{\strong{\code{DESeq2}}}{Fit the GLM defined in the function \code{\link{DESeq}}, which is the NB GLM with regulated estimation of the overdispersion parameters.}

\item{\strong{\code{NB}}}{Fit the Negative Binomial (NB) GLM.}

\item{\strong{\code{Poisson}}}{Fit the Poisson GLM.}
}

By default, the DESeq2 GLMs are fitted on the data set with > 1 biological replicates for both the IP and input samples, the Poisson GLM will be fitted otherwise.}

\item{LFC_shrinkage}{a \code{character} for the method of emperical bayes shrinkage on log2FC, could be one of \code{c("apeglm", "ashr", "none")}; Default \code{= "apeglm"}.

see \code{\link{lfcShrink}} for more details.}

\item{...}{Optional arguments passed to \code{\link{DESeq}}}
}
\description{
\code{glmDM} perform inference and estimation on RNA differential modification log2FC.

GLMs with interactive design between dummy variables of IP/input and Treatment/control are fitted for each peaks/sites:

\deqn{log2(Q) = intercept + I(Treatment) + I(IP) + I(IP)*I(Treatment)}

The log2FC and the associated statistics are based on the coefficient estimate of the interactive term： \eqn{I(IP)*I(Treated)}.

Under default setting, the returned log2FC are the RR estimates with Couchey priors defined in \code{\link{apeglm}}.
}
\examples{

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)

aln <- scanMeripBAM(
bam_ip = c("IP_rep1.bam",
           "IP_rep2.bam",
           "IP_rep3.bam"),
bam_input = c("input_rep1.bam",
              "input_rep2.bam",
              "input_rep3.bam"),
bam_treated_ip = c("IP_treated_rep1.bam",
                   "IP_treated_rep2.bam"),
bam_treated_input = c("input_treated_rep1.bam",
                      "input_treated_rep2.bam"),
paired_end = TRUE
)

sep <- exomePeakCalling(merip_bams = aln,
                        txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
                        bsgenome = Hsapiens)

sep <- normalizeGC(sep)

sep <- glmDM(sep)


}
\seealso{
\code{\link{glmM}}
}
