% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MeripBamFileList-class}
\alias{MeripBamFileList-class}
\alias{MeripBamFileList}
\title{MeripBamFileList}
\description{
An object that summarizes the BAM files used in a MeRIP-Seq experiment.
}
