% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/normalizeGC.R
\docType{methods}
\name{normalizeGC}
\alias{normalizeGC}
\alias{normalizeGC}
\title{Estimation of the GC content normalization factors.}
\usage{
normalizeGC(sep, bsgenome = "hg19", txdb = "hg19", gene_annot = NULL,
  fragment_length = 100, binding_length = 25, feature = c("background",
  "all"), qtnorm = FALSE, effective_GC = FALSE)

\S4method{normalizeGC}{SummarizedExomePeak}(sep, bsgenome = NULL,
  txdb = NULL, gene_annot = NULL, fragment_length = 100,
  binding_length = 25, feature = c("background", "all"), qtnorm = TRUE,
  effective_GC = FALSE)
}
\arguments{
\item{sep}{a \code{summarizedExomePeak} object.}

\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence, or it could be the name of the reference genome recognized by \code{\link{getBSgenom}}.}

\item{txdb}{a \code{\link{TxDb}} object for the transcript annotation, or it could be the name of the reference genome recognized by \code{\link{makeTxDbFromUCSC}}.}

\item{fragment_length}{the expected fragment length of the sequencing library; Default 100.}

\item{binding_length}{the expected antibody binding length of IP; Default 25.}

\item{feature}{the features used in the GC effect estimation, can be "background" and "all".
If "all" is choosed, the GC effect function will be estimated using both the methylated and the background regions,
this choice will force the resulting modification signals independent of GC content, which could be more robust when the background is incorrectly estimated.
However,it is possibably biased if the modification level is biologically dependent on GC content;
Default "background".}

\item{qtnorm}{A logical indicating wheather perform quantile normalization after the GC effect estimation； default TRUE.
Quantile normalization will be conducted on IP and input samples seperately to account for the biological difference between the marginal distributions of IP and input.}

\item{effective_gc}{whether to calculate the weighted GC content by the probability of reads alignment; default FALSE.}
}
\value{
a \code{summarizedExomePeak} object containing the feature specific size factors.
}
\description{
This function estimates the feature specific size factors in order to correct the GC content artifacts,
the GC content biases can contributed to the following sources of variabilities in MeRIP-seq data:

1. Technical variablities between replicates.

2. Batch effects between different laboratories.

Using the default option, the GC normalization estimates the dependency between reads count and GC content only on the not methylated region,
This strategy can avoid the quantification biases for the modifications that biologically favor extream GC content.

The GC normalization can results in an improvement of accuracy for most published m6A-seq data,
and it is particullarly recommended if you want to compare the data between different laboratory conditions.
}
\details{
By default, the sequencing depth factor used is the column size factors obtained by \code{\link{estimateSeqDepth}},
if the column size factors are not provided, it will be estimated using the default method defined in \code{\link{estimateSeqDepth}}.
}
