% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEffectGC.R
\docType{methods}
\name{plotEffectGC}
\alias{plotEffectGC}
\alias{plotBetaGC}
\title{Plot the relationship between the effect sizes (GLM log2FC estimates) and GC content.}
\usage{
plotEffectGC(sep, bsgenome = NULL, txdb = NULL, save_pdf_prefix = NULL,
  fragment_length = 100, binding_length = 25, effective_GC = FALSE)

\S4method{plotEffectGC}{SummarizedExomePeak}(sep, bsgenome = NULL,
  txdb = NULL, save_pdf_prefix = NULL, fragment_length = 100,
  binding_length = 25, effective_GC = FALSE)
}
\arguments{
\item{bsgenome}{a \code{\link{BSgenome}} object for the genome sequence, it could be the name of the reference genome recognized by \code{\link{getBSgenom}}.}

\item{txdb}{a \code{\link{TxDb}} object for the transcript annotation, it could be the name of the reference genome recognized by \code{\link{makeTxDbFromUCSC}}}

\item{save_pdf_prefix}{a character, if provided, a pdf file with the given name will be saved under the current directory.}

\item{fragment_length}{the expected fragment length of the sequencing library; Default 100.}

\item{binding_length}{the expected antibody binding length of IP; Default 25.}

\item{effective_gc}{whether to calculate the weighted GC content by the probability of reads alignment; default FALSE.}
}
\value{
a ggplot object.
}
\description{
This function plot the scatter plot between GC content and the methylation / differential methylation levels of the modification sites.
}
\details{
By default, this function will generate the countour of the scatter plot, and a linear regression line indicating the trend between
GC content and log2 fold change or log2 odds ratio returned by DESeq2.
The significant changed methylation sites will be grouped and lebeled in different colours.
}
