% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExonLength.R
\docType{methods}
\name{plotExonLength}
\alias{plotExonLength}
\title{Plot the Distribution of the Length of Exons Overlapped by the RNA Modification Peaks/Sites}
\usage{
plotExonLength(sep, txdb = NULL, save_pdf_prefix = NULL,
  include_control_regions = TRUE, save_dir = ".")

\S4method{plotExonLength}{SummarizedExomePeak}(sep, txdb = NULL,
  save_pdf_prefix = NULL, include_control_regions = TRUE,
  save_dir = ".")
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeaks}} object.}

\item{txdb}{a \code{\link{TxDb}} object containing the transcript annotation.}

\item{save_pdf_prefix}{a \code{character} if provided, a pdf file with the given name will be saved under the current working directory.}

\item{include_control_regions}{a \code{logical} for whether to include the control regions or not; Default \code{= TRUE}.}

\item{save_dir}{a \code{character} for the directory to save the plot; Default \code{= "."}.}
}
\value{
a ggplot object
}
\description{
This function plot the distribution of the exon length for peaks containing exons.
}
\details{
If the SummarizedExomePeaks object contains LFC statistics, the significantly modified peaks
with IP to input log2FC > 0 and DESeq2 Wald test padj < .05 will be plotted .

If the SummarizedExomePeaks object contains interactive LFC statistics, both the hyper modification
and hypo modification peaks with DESeq2 Wald test p values < .05 will be plotted.
}
\examples{

\dontrun{
sep #a SummarizedExomePeak object
txdb #a TxDb object
plotExonLength(sep,txdb)
}

}
