% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSizeFactors.R
\docType{methods}
\name{plotSizeFactors-methods}
\alias{plotSizeFactors-methods}
\alias{plotSizeFactors,SummarizedExomePeak-method}
\alias{plotSizeFactors}
\title{Method plotSizeFactors}
\usage{
plotSizeFactors(sep)

\S4method{plotSizeFactors}{SummarizedExomePeak}(sep)
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeak}} object.}
}
\description{
Method plotSizeFactors

plot the size factors using different strategies.
}
