% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_five_POS.R
\name{reads_five_POS}
\alias{reads_five_POS}
\title{extract reads five prime POS.}
\usage{
reads_five_POS(reads, width = 1, fix = "start", ...)
}
\arguments{
\item{reads}{a GAlignmentList object.}
}
\description{
extract reads five prime POS.
}
