% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_five_POS_rev.R
\name{reads_five_POS_rev}
\alias{reads_five_POS_rev}
\title{extract reads five prime POS and reverse its strand.}
\usage{
reads_five_POS_rev(reads, width = 1, fix = "start", ...)
}
\arguments{
\item{reads}{a GAlignmentList object.}
}
\description{
extract reads five prime POS and reverse its strand.
}
