% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Results.R
\docType{methods}
\name{Results}
\alias{Results}
\title{Report the (Differential) Modification Peaks/Sites and their associated LFC Statistics}
\usage{
Results(sep, cut_off_pvalue = NULL, cut_off_padj = 0.05,
  cut_off_log2FC = 0, min_num_of_positive = 30,
  expected_direction = "both", inhibit_filter = FALSE,
  table_style = c("bed", "granges"))

\S4method{Results}{SummarizedExomePeak}(sep, cut_off_pvalue = NULL,
  cut_off_padj = 0.05, cut_off_log2FC = 0, min_num_of_positive = 30,
  expected_direction = "both", inhibit_filter = TRUE,
  table_style = c("bed", "granges"))
}
\arguments{
\item{sep}{a \code{\link{SummarizedExomePeak}} object.}

\item{cut_off_pvalue}{a \code{numeric} value for the p value cutoff in the exported result; Default \code{= NULL}.}

\item{cut_off_padj}{a \code{numeric} value for the adjusted p value cutoff in the exported result; Default \code{= 0.05}.}

\item{cut_off_log2FC}{a \code{numeric} value for the log2 fold change (LFC) cutoff of the exported result,
only the sites with abs(LFC) larger than this value are kept; Default \code{= 0}.}

\item{min_num_of_positive}{a \code{numeric} value for the minimum number of reported sites.
If the number of remaining sites is less than this number after the filter, additional sites will be reported by the increasing order of the p value to meet this number.}

\item{expected_direction}{a \code{character} for the expected direction of the differential modification, could be one in \code{c("hyper", "hypo", "both")}.

\describe{
 \item{\strong{\code{hyper}}}{
 only report the peaks/sites with interactive LFC > 0.
 }

 \item{\strong{\code{hypo}}}{
 only report the peaks/sites with interactive LFC < 0.
 }

 \item{\strong{\code{both}}}{
 report the peaks/sites in both directions.
 }
}

This argument is useful when the treated group involves the perturbation of a known writer or eraser protein; Default "both".}

\item{inhibit_filter}{a \code{logical} for whether to remove all the filters, this option is useful when quantification on single based site annotation; Default \code{= FALSE}.}

\item{table_style}{a \code{character} for the style of the tsv table being exported, could be one in \code{c("bed","granges")}.

\describe{
 \item{\strong{\code{bed}}}{
 The genomic locations in the table are represented by BEDgraph style.
 }

 \item{\strong{\code{granges}}}{
 The genomic locations in the table are represented by GRanges style.
 }
}}
}
\value{
a \code{data.frame} containing the genomic locations of modification peaks/sites, gene ids, and their statistics.
}
\description{
Report the (Differential) Modification Peaks/Sites and their associated LFC Statistics
}
\examples{

library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(BSgenome.Hsapiens.UCSC.hg19)

aln <- scanMeripBAM(
bam_ip = c("IP_rep1.bam",
           "IP_rep2.bam",
           "IP_rep3.bam"),
bam_input = c("input_rep1.bam",
              "input_rep2.bam",
              "input_rep3.bam"),
paired_end = TRUE
)

sep <- exomePeakCalling(merip_bams = aln,
                        txdb = TxDb.Hsapiens.UCSC.hg19.knownGene,
                        bsgenome = Hsapiens)

sep <- normalizeGC(sep)

sep <- glmM(sep)

exportResults(sep)

}
